remove a bunch of 'as usize'

+++ crates/aptos-dkg/src/pvss/chunky/chunked_elgamal.rs
@@ -363,7 +363,7 @@ where
     F: PrimeField, // need `PrimeField` here because of `sample_field_element()`
     R: rand_core::RngCore + rand_core::CryptoRng,
 {
-    let mut r_vals = vec![F::zero(); num_chunks as usize];
+    let mut r_vals = vec![F::zero(); num_chunks];
     let mut remaining = *target_sum;
     let radix_f = F::from(radix);
     let mut cur_base = radix_f;
